+++ src/main/java/com/tarasov/footballproject/repositores/TeamRepository.java
@@ -1,8 +1,5 @@
 package com.tarasov.footballproject.repositores;
 
-import com.tarasov.footballproject.dto.get.GetFullTeamInfoDTO;
-import com.tarasov.footballproject.entities.City;
-import com.tarasov.footballproject.entities.Stadium;
 import com.tarasov.footballproject.entities.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -11,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 
     @Query("select t from Team t where t.teamName = ?1")
     Team findTeamByTeamNameIgnoreCase(String name);
